requirements.txt
appdirs==1.4.3
attrs==19.1.0
boto==2.49.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Collectfast==0.2.0
contextlib2==0.5.5
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.13
django-allauth==0.39.1
django-appconf==1.0.3
django-autoslug==1.9.4
django-bootstrap3==11.0.0
django-braces==1.13.0
django-configurations==2.1
django-debug-toolbar==2.0
django-dotenv==1.4.2
django-extensions==2.1.9
django-fsm==2.6.1
django-fsm-log==1.6.2
django-ical==1.5
django-markdown-deux==1.0.5
django-pylibmc==0.6.1
django-storages==1.7.1
django-taggit==1.1.0
ecdsa==0.13.2
Fabric==1.8.3
gunicorn==18.0
icalendar==4.0.3
idna==2.8
markdown2==2.3.8
newrelic==2.18.1.15
oauthlib==3.0.1
pinax-notifications==5.0.3
psycopg2==2.8.3
pycrypto==2.6.1
pylibmc==1.6.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
raven==6.10.0
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tweepy==3.7.0
Unidecode==0.4.14
urllib3==1.25.3
Werkzeug==0.15.4
